requirements.txt
discord.py
aiosqlite
python-dotenv
pytest
pytest-asyncio